factor: 3주차 수요일 피드백 반영
