--- app/src/main/java/com/example/it/issuetracker/data/datasource/MilestoneDataSourceDefaultImpl.kt
@@ -1,54 +1,0 @@
-package com.example.it.issuetracker.data.datasource
-
-import android.util.Log
-import com.example.it.issuetracker.data.dto.AddMilestoneDto
-import com.example.it.issuetracker.data.dto.MilestoneDto
-import com.example.it.issuetracker.data.network.IssueTrackerService
-
-private const val TAG = "MilestoneDataSourceImpl"
-
-class MilestoneDataSourceDefaultImpl(
-    private val issueTrackerApi: IssueTrackerService,
-) : MilestoneDataSource {
-
-    override suspend fun getMilestoneInfoList(): List<MilestoneDto> {
-        try {
-            return issueTrackerApi.getMilestoneInfoList()
-        } catch (e: Exception) {
-            Log.e(TAG, e.message, e)
-            error(e)
-        }
-    }
-
-    override suspend fun addMilestone(addMilestoneDto: AddMilestoneDto) {
-        try {
-            issueTrackerApi.addMilestone(addMilestoneDto)
-        } catch (e: Exception) {
-            Log.e(TAG, e.message, e)
-            error(e)
-        }
-    }
-
-    override suspend fun editMilestone(milestoneDto: MilestoneDto) {
-        try {
-            val editMilestoneDto = AddMilestoneDto(
-                milestoneDto.title,
-                milestoneDto.description ?: "",
-                milestoneDto.deadline ?: ""
-            )
-            issueTrackerApi.editMilestoneInfo(milestoneDto.id!!, editMilestoneDto)
-        } catch (e: Exception) {
-            Log.e(TAG, e.message, e)
-            error(e)
-        }
-    }
-
-    override suspend fun deleteMilestone(id: Long) {
-        try {
-            issueTrackerApi.deleteMilestone(id)
-        } catch (e: Exception) {
-            Log.e(TAG, e.message, e)
-            error(e)
-        }
-    }
-}
