+++ app/src/main/java/com/example/it/issuetracker/di/sharedPreferenceModule.kt
@@ -1,6 +1,6 @@
 package com.example.it.issuetracker.di
 
-import com.example.it.issuetracker.data.datasource.UserSharedPrefDataSource
+import com.example.it.issuetracker.data.datasource.local.UserSharedPrefDataSource
 import org.koin.android.ext.koin.androidContext
 import org.koin.dsl.module
