+++ app/src/main/java/com/example/it/issuetracker/presentation/common/Constants.kt
@@ -15,4 +15,9 @@ object Constants {
 
     // Issue 추가 및 수정 부분에서 번들로 받기 위한 Key값 정의
     const val ISSUE_BUNDLE_KEY = "issue_bundle"
+
+    // Swipe
+    const val PERCENT = 10
+    const val RATIO = 3
+    const val REVERSE = -1f
 }
