+++ app/src/main/java/com/example/it/issuetracker/presentation/main/mypage/MyPageViewModel.kt
@@ -3,7 +3,7 @@ package com.example.it.issuetracker.presentation.main.mypage
 import androidx.lifecycle.ViewModel
 import androidx.lifecycle.viewModelScope
 import com.example.it.issuetracker.R
-import com.example.it.issuetracker.data.datasource.UserSharedPrefDataSource
+import com.example.it.issuetracker.data.datasource.local.UserSharedPrefDataSource
 import com.example.it.issuetracker.domain.repository.LoginRepository
 import com.example.it.issuetracker.presentation.common.Constants
 import com.example.it.issuetracker.presentation.login.LoginOption
